+++ test/structure-optimizer.spec.js
@@ -43,6 +43,23 @@ describe("AST structure optimizer", () => {
     });
   });
 
+  it("changes `a | b | [ b ]` in ast to choice with skip without duplicates", () => {
+    const text = "foo = a | b | [ b ];";
+    const result = parser.parse(text);
+    const inputDefinition = result[0].definition;
+    expect(inputDefinition).to.eql({
+      choice: [
+        { nonTerminal: "a" },
+        { nonTerminal: "b" },
+        { optional: { nonTerminal: "b" } }
+      ]
+    });
+    const optimizedDefinition = optimizeProduction(inputDefinition);
+    expect(optimizedDefinition).to.eql({
+      choice: [{ skip: true }, { nonTerminal: "a" }, { nonTerminal: "b" }]
+    });
+  });
+
   it("changes `a | { b }` in ast to choice with skip, a and b+", () => {
     const text = "foo = a | { b };";
     const result = parser.parse(text);
@@ -146,7 +163,8 @@ describe("AST structure optimizer", () => {
     });
     const optimizedDefinition = optimizeProduction(inputDefinition);
     expect(optimizedDefinition).to.eql({
-      sequence: [{ repetition: { nonTerminal: "a" }, skippable: false }]
+      repetition: { nonTerminal: "a" },
+      skippable: false
     });
   });
 
@@ -162,7 +180,8 @@ describe("AST structure optimizer", () => {
     });
     const optimizedDefinition = optimizeProduction(inputDefinition);
     expect(optimizedDefinition).to.eql({
-      sequence: [{ repetition: { nonTerminal: "a" }, skippable: false }]
+      repetition: { nonTerminal: "a" },
+      skippable: false
     });
   });
 
@@ -203,6 +222,155 @@ describe("AST structure optimizer", () => {
               sequence: [{ nonTerminal: "e" }, { nonTerminal: "d" }]
             },
             skippable: false
+          }
+        ]
+      });
+    }
+  );
+
+  it(
+    "changes `a, b, c, { b, c }` in ast to " + "`a, (b, c)+` without repeater",
+    () => {
+      const text = "foo = a, b, c, { b, c };";
+      const result = parser.parse(text);
+      const inputDefinition = result[0].definition;
+      expect(inputDefinition).to.eql({
+        sequence: [
+          { nonTerminal: "a" },
+          { nonTerminal: "b" },
+          { nonTerminal: "c" },
+          {
+            repetition: {
+              sequence: [{ nonTerminal: "b" }, { nonTerminal: "c" }]
+            },
+            skippable: true
+          }
+        ]
+      });
+      const optimizedDefinition = optimizeProduction(inputDefinition);
+      expect(optimizedDefinition).to.eql({
+        sequence: [
+          { nonTerminal: "a" },
+          {
+            repetition: {
+              sequence: [{ nonTerminal: "b" }, { nonTerminal: "c" }]
+            },
+            skippable: false
+          }
+        ]
+      });
+    }
+  );
+
+  it("changes `a | a` into " + "`a`", () => {
+    const text = "foo = a | a;";
+    const result = parser.parse(text);
+    const inputDefinition = result[0].definition;
+    expect(inputDefinition).to.eql({
+      choice: [{ nonTerminal: "a" }, { nonTerminal: "a" }]
+    });
+    const optimizedDefinition = optimizeProduction(inputDefinition);
+    expect(optimizedDefinition).to.eql({ nonTerminal: "a" });
+  });
+
+  it("changes `a, b, c | a` into " + "`a, [ b, c ]`", () => {
+    const text = "foo = a, b, c | a;";
+    const result = parser.parse(text);
+    const inputDefinition = result[0].definition;
+    expect(inputDefinition).to.eql({
+      choice: [
+        {
+          sequence: [
+            { nonTerminal: "a" },
+            { nonTerminal: "b" },
+            { nonTerminal: "c" }
+          ]
+        },
+        { nonTerminal: "a" }
+      ]
+    });
+    const optimizedDefinition = optimizeProduction(inputDefinition);
+    expect(optimizedDefinition).to.eql({
+      sequence: [
+        { nonTerminal: "a" },
+        {
+          optional: {
+            sequence: [{ nonTerminal: "b" }, { nonTerminal: "c" }]
+          }
+        }
+      ]
+    });
+  });
+
+  it(
+    "changes `a, b, c | a, d, e | a, f, e` into " +
+      "`a, ( b, c | ( d | f ), e )`",
+    () => {
+      const text = "foo = x, y, z | a, b, c | a, b, d, e | a, b, f, e;";
+      const result = parser.parse(text);
+      const inputDefinition = result[0].definition;
+      expect(inputDefinition).to.eql({
+        choice: [
+          {
+            sequence: [
+              { nonTerminal: "x" },
+              { nonTerminal: "y" },
+              { nonTerminal: "z" }
+            ]
+          },
+          {
+            sequence: [
+              { nonTerminal: "a" },
+              { nonTerminal: "b" },
+              { nonTerminal: "c" }
+            ]
+          },
+          {
+            sequence: [
+              { nonTerminal: "a" },
+              { nonTerminal: "b" },
+              { nonTerminal: "d" },
+              { nonTerminal: "e" }
+            ]
+          },
+          {
+            sequence: [
+              { nonTerminal: "a" },
+              { nonTerminal: "b" },
+              { nonTerminal: "f" },
+              { nonTerminal: "e" }
+            ]
+          }
+        ]
+      });
+      const optimizedDefinition = optimizeProduction(inputDefinition);
+      expect(optimizedDefinition).to.eql({
+        choice: [
+          {
+            sequence: [
+              { nonTerminal: "a" },
+              { nonTerminal: "b" },
+              {
+                choice: [
+                  {
+                    sequence: [
+                      {
+                        choice: [{ nonTerminal: "d" }, { nonTerminal: "f" }]
+                      },
+                      { nonTerminal: "e" }
+                    ]
+                  },
+                  { nonTerminal: "c" }
+                ]
+              }
+            ]
+          },
+          {
+            sequence: [
+              { nonTerminal: "x" },
+              { nonTerminal: "y" },
+              { nonTerminal: "z" }
+            ]
           }
         ]
       });